Improve security through avoiding env leaks
This passes the environment, where the PAT is located, through an
environment file instead of a command line argument. This is a security
improvement, as it avoids leaking the PAT through the process list. The
environment file is automatically removed before the runner process is
created, ensuring that it cannot be accessed from within workflows.