Configure from installation copies
When starting several runners, the same token will be returned. In order
to be able to run several runners, they need to be configured from
different working directories, and these directories need to contain an
entire copy of the runner installation.

Arrange for two sub-directories to be created in the working directory:
one called `runner` will contain the copy of the installation tree,
another called `work` will be where the runner will be "working", e.g.
checkout repositories, etc.