Improve handshaking between processes
A .tkn file, containing the runner registration token is created inside
the microVM. The main orchestrator will wait for the first such token
file for a given runner loop before continuing iterating to the next
runner loop. This slows down startup time slightly, but provides
improved determinism.

Also, when an exit signal is caught, a file with the extension .brk is
created with a secret. This file and the content of the secret are
verified in the runner loop, and, when present and matching, the loop is
stopped instead of a new runner being created.