Handle termination signals
Implement graceful shutdown of all the runners on termination signals
received by the orchestrator. In order to avoid ctrl-c from being sent
down the entire tree, a separate process group is created by the runner
loop: this prevents abrupt termination of the microVM. When termination
is received, a termination file (with the same secret as for the break
file) is created. The implementation will react to the termination file
and kill and unregister the runner.

In order for all termination to be handled in the background, all
processes are started in the background and waited for. The library is
enhanced with a number of functions to handle processes, these should be
portable to macOS.