No separate copy when ephemeral runner is used
When the runner is known to be ephemeral, do not copy the content of the
runner unpacked distribution to the work directory. Instead, run from
there directly and set the RUNNER_BINROOT variable to point to the
installation directory. Ensure that the content of the variable is used
for deciding where to run the runner from and fix directory tree
ownership prior to starting the runner.

Rename the label to krun, instead of krunvm. This is to highlight the
underlying technology used, rather than how the runners are created.
This is a breaking change, but it is not expected to be used in
production yet.