podman + krun Support (#25)
* Add dev note for image cleanup

* Add generic microvm API lib and use it

Add a generic API to create microVMs in an "agnostic" way, and migrate
to the new API in the orchestrator.sh and runner.sh scripts. While the
API has code for podman, this update only focuses on migrating the
existing krunvm-based code to the new API.

The new API is implemented in lib/microvm.sh, and provides a set of
docker/podman lookalike commands to manage microVMs. The function
microvm_runtime can be used to pinpoint or auto-detect the runtime to
use depending on commands available at the host. The API is **not** able
to handle several krunvm-based microVMs at the same time.

* Add support for multiple krunvm VMs

Add support for multiple krunvm-based VMs in the API. This uses a
temporary directory that will store the mappings between the name of the
VM and the PID of the underlying (top) process. The directory should be
removed through microvm_cleanup upon process exit.

* Enforce mount options for podman

* Add support for other podman runtimes

While the podman runtime is still enforced, the script now supports
several other runtimes. The KRUNVM_RUNNER_RUNTIME variable should
contain the word podman, followed by the plus sign, followed by the
runtime to use, e.g. podman+krun will pass the value of krun to the
--runtime option of podman create/run.

* Add mount option only if not present

* Add documentation

* Add ref to installation on old Fedoras

* Enhance cleanup section

* Check command is passed

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>

* Fix proper variable

* Add error when no runtime found

* Fix spelling in comment

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>